+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Unreleased
 ----------
-n/a
+**Added**
+**Changed**
+**Fixed**
+  ConnectionWrapper.execute does not pass the argument ``arguments`` to
+  the underlying cursor's execute method if ``arguments`` is None. Some drivers
+  raise an Error if None is passed, some don't.
+
 
 Version 2.5
 -----------